+++ Cargo.toml
@@ -44,6 +44,7 @@ fmi-test-data = { path = "fmi-test-data", version = "0.5.0" }
 itertools = "0.14"
 libloading = "0.8"
 log = "0.4"
+paste = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 tempfile = "3.1"
 test-log = { version = "0.2", features = ["trace"] }
