+++ fmi-export/Cargo.toml
@@ -13,16 +13,14 @@ include.workspace = true
 [features]
 default = ["fmi3"]
 ## Enable support for FMI 3.0
-fmi3 = ["fmi/fmi3", "fmi-sys/fmi3"]
+fmi3 = ["fmi/fmi3"]
 
 [dependencies]
 chrono = { workspace = true }
 document-features = { workspace = true }
 fmi = { workspace = true }
-fmi-sys = { workspace = true }
 fmi-export-derive = { workspace = true }
-log = { workspace = true }
-#serde = { workspace = true }
+paste = { workspace = true }
 thiserror = { workspace = true }
 
 [dev-dependencies]
